pyproject.toml
[tool.poetry]
name = "hardsploit"
version = "0.2.0"
description = "API for Hardsploit HW access"
authors = ["PERIES Mickaël <m.peries@serma.com>", "DUMAN Mazlum, <m.duman@serma.com>"]
license = "LGPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/serma-safety-security/hardsploit"
documentation = "https://github.com/serma-safety-security/hardsploit/wiki"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]
packages = [
    {include = "hardsploit"}
]
include = [ {path = "data/50-Hardsploit.rules" }, { path = "examples/" }]
exclude = [".gitignore"]

[tool.poetry.dependencies]
python = ">=3.8"
colorama = "^0.4.6"
pyusb = "^1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/serma-safety-security/hardsploit/issues"

